Move performDeleteOnRelations method into HasRelationship concern
This method is strictly a relationship method and should be grouped with the other relationship methods.
