+++ src/Database/Model.php
@@ -992,50 +992,7 @@ class Model extends EloquentModel implements ModelInterface
     {
         $this->performDeleteOnRelations();
 
-        $this->setKeysForSaveQuery($this->newQueryWithoutScopes())->delete();
-    }
-
-    /**
-     * Locates relations with delete flag and cascades the delete event.
-     * For pivot relations, detach the pivot record unless the detach flag is false.
-     * @return void
-     */
-    protected function performDeleteOnRelations()
-    {
-        $definitions = $this->getRelationDefinitions();
-        foreach ($definitions as $type => $relations) {
-            /*
-             * Hard 'delete' definition
-             */
-            foreach ($relations as $name => $options) {
-                if (in_array($type, ['belongsToMany', 'morphToMany', 'morphedByMany'])) {
-                    // we want to remove the pivot record, not the actual relation record
-                    if (Arr::get($options, 'detach', true)) {
-                        $this->{$name}()->detach();
-                    }
-                } elseif (in_array($type, ['belongsTo', 'hasOneThrough', 'hasManyThrough', 'morphTo'])) {
-                    // the model does not own the related record, we should not remove it.
-                    continue;
-                } elseif (in_array($type, ['attachOne', 'attachMany', 'hasOne', 'hasMany', 'morphOne', 'morphMany'])) {
-                    if (!Arr::get($options, 'delete', false)) {
-                        continue;
-                    }
-
-                    // Attempt to load the related record(s)
-                    if (!$relation = $this->{$name}) {
-                        continue;
-                    }
-
-                    if ($relation instanceof EloquentModel) {
-                        $relation->forceDelete();
-                    } elseif ($relation instanceof CollectionBase) {
-                        $relation->each(function ($model) {
-                            $model->forceDelete();
-                        });
-                    }
-                }
-            }
-        }
+        parent::performDeleteOnModel();
     }
 
     //
